Started Crew Page and finished it
"Started working with Crew Page. What was challenging was to find a way for the main layout to be the same as in destination page but adapt the rest to how this page should look. Was a bit confusing on how I should map through the images and whether or not I should have a general image style and reuse it.."